Refactor: 상수 분리
