+++ src/js/constants/productManageMenu.js
@@ -3,8 +3,3 @@ export const PRODUCT = {
   MIN_PRICE: 100,
   DIVISIBLE_UNIT: 10,
 };
-
-export const CHARGE = {
-  MIN_PRICE: 100,
-  DIVISIBLE_UNIT: 10,
-};
