+++ src/js/constants/vendingMachineManageMenu.js
@@ -1,0 +1,21 @@
+export const CHARGE = {
+  INIT: 0,
+  MIN_PRICE: 100,
+  DIVISIBLE_UNIT: 10,
+};
+
+export const COIN_500 = '500';
+export const COIN_100 = '100';
+export const COIN_50 = '50';
+export const COIN_10 = '10';
+
+export const COINS = {
+  INIT: {
+    500: 0,
+    100: 0,
+    50: 0,
+    10: 0,
+  },
+
+  UNITS: [COIN_500, COIN_100, COIN_50, COIN_10],
+};
